Prevent unsafe run in absence of GNU coreutils
It has been found to be unsafe to run bitpocket if the coreutils
(ls/cp/mv/mkdir/sed/...) in the system are not the GNU coreutils.
In particular, running bitpocket on systems with BSD sed can
cause data loss.

The permanent resolution is to make sure that bitpocket will
work with any version of these tools. In the meantime, we
disable the use of bitpocket on systems where it is not safe
to use (and will not work).