reword: sync: Fix dropping of local files
Fixes an issue where a file does not appear as new on the local side but does
not exist on the remote side. When the sync is performed the local file is
removed. If it is re-created locally, it will again not appear as new since the
previous sync marked it as existing before running the sync.

This fix suggests a file listing should be captured of the remote side and
compared with the file listing on the local side. This will help assess files
which appear to have been deleted remotely but still exist locally, as well as
files which exist locally only but do not exist remotely. This comparison can
be performed without comparing current and previous file listings.

It also incorporates several speed improvements, especially where using rsync
to synchronize extended attributes.