Fix push of remotely-deleted files
This adjusts the matching mechanism to create a list of locally added and
deleted files, and remotely deleted files. Then, when the push and pull are
run, locally added files are excluded from the pull and remotely deleted
files are excluded from the push.

This fixes the issue where remotely deleted files are not pulled, but then,
because they were not excluded, were also pushed back to the remote server.
Therefore, if files were removed remotely, they would appear again during
the sync.