Disambiguate files added in parallel with the sync from those pulled
This removes the usage of the `added-prev` file which included both files
fetched from the remote server in the `pull` process as well as files
created locally during the sync. Instead, it adds and removes filenames of
files added or deleted during the `pull` part of the sync to the `tree-prev`
file. This allows the files which were added in parallel with the `pull` to
be naturally included in the next run of the sync process.

This patch also removes another invocation of `rsync` to collect a local
snapshot of the file system after the sync has completed.