Add support for remote backups
Use `rsync` to do the backing up

rsync has a --backup option which will accomplish what the backup_inline
function is accomplishing. However, it will add the ability to do remote
backups without doing remote shell magic.

Coincidently, this removes the parallel rsync used to do the pull(), which
may reduce the CPU and disk overhead slightly.

This allows the remote host to cooperate with the PUSH process and create
backups on the remote host itself, for files which will be overwritten by the
PUSH.

Remote backups is an optional feature, disabled by default, and can be
enabled with the REMOTE_BACKUPS option in the configuration file.