pull: Fix pull of recursive dirs
If a folder is received with the --files-from parameter, and the folder name
ends with a trailing slash, then rsync will copy the contents of the folder
recursively. This is not intended in the sync process. Instead, only the stat
information on the folder should be pulled. The files in the folders will be
communicated via separate line items.