pull: Only delete local files if removed remotely
This patch adds a check before local deletion to inspect if a file to be
deleted was marked as remotely deleted.

This should help a situation where, when a sync runs for a long time (more than
a few seconds), it becomes likely that a new file will be created on the local
or remote system. Such a file could be removed by the sync if it sorts
alphabetically after large files holding the sync up. When the new file is
encountered, it is removed because it was not protected by the exclude list in
`local-add-del`.