push: Protect remotely-created files in parallel
This is similar to the code to protect locally-created files in parallel
with a long-running sync. Since files which show up after the sync takes a
directory snapshot are not protected, they could be deleted. This protects
all the remote files except for the locally-deleted files. So, the only
files removed in the push are the ones which were removed locally since the
previous sync.