sync: Add optional check for mounted mountpoint(s)
If the systems being synced by bitpocket are syncing a special mounted
filesystem, then a check can be added to ensure the filesystem is mounted
before the sync will begin. This allows for bitpocket to gracefully abort the
sync instead of deleting all the local or remote data because of the missing
mount (making the sync target appear empty like all the files were removed).