sync: Fix pull clobbering local mode changes
This patch fixes the pull process so that local changes to a file's mode (file
type such as transition from link to file or visa versa and permissions) are
preserved if changed locally between invocations of bitpocket.

This change is represented by changes in the format of the `tree-prev` and
`tree-current` files to include the file's mode.

This changes gracefully upgrades the state file from previous versions. Since
the mode is added to the end of the filenames, its absense does not prevent the
comparison of current and previous file listings.

In this version, the mode is added to the beginning of the line in the state
file and preceeded by a colon (:). This helps differentiate between the
previous version of the state file which began with a leading slash (/).