analyse: Speed up local change detection
The previous version did not work correctly when transitioning from the old
state format to the new state format. This patch takes the STATE_VERSION into
consideration and properly detects changes when transitioning.

Instead of resorting the current and previous trees, use the files sorted on
the 12th column and add a false-positive detection. This seems to be accurate
and much faster for large trees (about 20x for 300k files).