sync: Add transaction support
In this mode, the files synced in both the pull and push phases are not put
into place until the completion of the phase. Instead, the updated files are
staged into a `.bitpocket/staging` folder. Once all the changes are staged,
they are moved into place. If the phase is interrupted, perhaps due to network
connection quality, then no changes are reflected. This is especially useful
when the changes being synchronized include software or data whose changes
should be reflected as a whole.

This has the caveat of requiring a bit of extra space on the source and
destination systems as extra space for all changes is required. However, it has
the caveat that, if interrupted, the staged partial file can be used when
resuming the transfer. This can greatly speed syncs over shaky network
connections.