perf: Abort sync if local and remote states are in sync
This will take an md5sum of the rsync output of both the local and
remote trees as they are output by rsync itself. This includes the mode,
modify time, full file path, and size. If all attributes of all files
match, then the md5 of the rsync listing stream will also match. With
this patch, the sync will exit early if it determines that neither a
pull nor a push is required.